CRUD Changes

+++ MultiThreading2/MultiThreading2/DeleteFile.cs
@@ -12,36 +12,57 @@ namespace MultiThreading2
     {
         public void textDelete()
         {
+            int count = 0;
             Thread.Sleep(2000);
             string textFile = @"C:\HPE-Files\MultiThreading\Text";
             string[] files = Directory.GetFiles(textFile);
             foreach(string text in files)
             {
-                File.Delete(text);
+                if (text.Length > 5000)
+                {
+
+
+                    File.Delete(text);
+                    count = count + 1;
+                }
                 Console.WriteLine("The Text File from Text folder is deleted");
             }
         }
 
         public void imageDelete()
         {
+            int count = 0;
             Thread.Sleep(3000);
             string imageFile = @"C:\HPE-Files\MultiThreading\Images";
             string[] allfiles = Directory.GetFiles(imageFile);
             foreach(string image in allfiles)
             {
-                File.Delete(image);
+                if (image.Length > 100000)
+                {
+
+                    File.Delete(image);
+                    count = count + 1;
+                    
+                }
                 Console.WriteLine("The Image File from Image Folder is deleted ");
             }
         }
 
         public void videoDelete()
         {
+            int count = 0;
             Thread.Sleep(4000);
             string videoFile = @"C:\HPE-Files\MultiThreading\VIdeo";
             string[] files = Directory.GetFiles(videoFile);
             foreach(string video in files)
             {
-                File.Delete(video);
+                if (video.Length > 1000000)
+                {
+
+
+                    File.Delete(video);
+                    count = count + 1;
+                }
                 Console.WriteLine("The video File from Video Folder is deleted");
             }
         }
